+++ src/main/java/com/mamouros/backend/ecoIsland/EcoIslandRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EcoIslandRepository extends CrudRepository<EcoIsland, Integer> {
+public interface EcoIslandRepository extends CrudRepository<EcoIsland, Long> {
 }
