+++ src/main/java/com/mamouros/backend/exceptions/IslandNotFoundException.java
@@ -2,7 +2,7 @@ package com.mamouros.backend.exceptions;
 
 public class IslandNotFoundException extends RuntimeException{
 
-    public IslandNotFoundException(Integer id){
+    public IslandNotFoundException(Long id){
         super("Could not find EcoIsland " + id);
     }
 }
