+++ src/main/java/com/mamouros/backend/exceptions/ReportNotFoundException.java
@@ -2,7 +2,7 @@ package com.mamouros.backend.exceptions;
 
 public class ReportNotFoundException extends RuntimeException{
 
-    public ReportNotFoundException(Integer id){
+    public ReportNotFoundException(Long id){
         super("Could not find Report" + id);
     }
 }
